+++ engine/client/ref_common.c
@@ -334,10 +334,6 @@ static ref_api_t gEngfuncs =
 	COM_FreeLibrary,
 	COM_GetProcAddress,
 
-	FS_LoadFile,
-	FS_FileExists,
-	FS_AllowDirectPaths,
-
 	R_Init_Video_,
 	R_Free_Video,
 
@@ -383,7 +379,9 @@ static ref_api_t gEngfuncs =
 
 	pfnDrawNormalTriangles,
 	pfnDrawTransparentTriangles,
-	&clgame.drawFuncs
+	&clgame.drawFuncs,
+
+	&g_fsapi,
 };
 
 static void R_UnloadProgs( void )
