+++ engine/wscript
@@ -166,7 +166,7 @@ def build(bld):
 			'client/vgui/*.c',
 			'client/avi/*.c'])
 
-	includes = ['common', 'server', 'client', 'client/vgui', 'tests', '.', '../public', '../common', '../pm_shared' ]
+	includes = ['common', 'server', 'client', 'client/vgui', 'tests', '.', '../public', '../common', '../filesystem', '../pm_shared' ]
 
 	if bld.env.SINGLE_BINARY:
 		install_path = bld.env.BINDIR
