+++ filesystem/wscript
@@ -1,0 +1,17 @@
+#!/usr/bin/env python
+
+def options(opt):
+	pass
+
+def configure(conf):
+	if conf.env.cxxshlib_PATTERN.startswith('lib'):
+		conf.env.cxxshlib_PATTERN = conf.env.cxxshlib_PATTERN[3:]
+
+def build(bld):
+	bld.shlib(target = 'filesystem_stdio',
+		features = 'c',
+		source = bld.path.ant_glob(['*.c']),
+		includes = ['.', '../common', '../public', '../engine'],
+		use = ['public'],
+		install_path = bld.env.LIBDIR,
+		subsystem = bld.env.MSVC_SUBSYSTEM)
