+++ public/crtlib.c
@@ -821,6 +821,23 @@ void COM_RemoveLineFeed( char *str )
 
 /*
 ============
+COM_FixSlashes
+
+Changes all '/' characters into '\' characters, in place.
+============
+*/
+void COM_FixSlashes( char *pname )
+{
+	while( *pname )
+	{
+		if( *pname == '\\' )
+			*pname = '/';
+		pname++;
+	}
+}
+
+/*
+============
 COM_PathSlashFix
 ============
 */
