+++ ref_soft/wscript
@@ -30,6 +30,7 @@ def build(bld):
 	source = bld.path.ant_glob(['*.c'])
 
 	includes = ['.',
+		'../filesystem',
 		'../engine',
 		'../engine/common',
 		'../engine/server',
